--- src/components/scss/LoginModal2.scss
@@ -1,155 +1,0 @@
-.SignUp {
-  display: none;
-  position: fixed;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  z-index: 99;
-  background-color: rgba(0, 0, 0, 0.6);
-  cursor: auto;
-  color:#ffb963;
-}
-.SignUp button {
-  outline: none;
-  cursor: pointer;
-  border: 0;
-}
-.header {
-  text-align: center;
-}
-
-.header-title {
-  font-size:3rem;
-  font-family: hopang;
-}
-.signinIcon {
-  width:15%;
-  margin:2em 0 1em 2em;
-}
-.SignUp > section {
-  width: 25%;
-  height: 70%;
-  margin: 0 auto;
-  border-radius: 0.3rem;
-  background-color: #fff;
-  /* 팝업이 열릴때 스르륵 열리는 효과 */
-  animation: modal-show 0.3s;
-  overflow: hidden;
-}
-.close {
-  float: right;
-  font-size: 1.5em;
-  outline: none;
-  cursor: pointer;
-  margin-right: 0.2em;
-}
-.SignUp > section > main {
-  padding: 16px;
-}
-.loginId {
-  margin-top: 30px;
-  border-radius: 2px;
-  width: 100%;
-  height: 40px;
-  border: 1px solid #e5e5e5;
-  padding: 9px 12px;
-  outline: none;
-  box-sizing: border-box;
-  font-size:0.85em;
-}
-input::placeholder {
-  color: #999999;
-}
-.loginPw {
-  margin-top: 15px;
-  border-radius: 2px;
-  width: 100%;
-  height: 40px;
-  border: 1px solid #e5e5e5;
-  padding: 9px 12px;
-  outline: none;
-  box-sizing: border-box;
-  font-size:0.85em;
-}
-.loginMid {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  .autoLogin {
-    font-size: 12px;
-    color: #8d8d8d;
-    line-height: 3;
-  }
-}
-.loginBtn {
-  height: 40px;
-  width: 100%;
-  font-size: 14px;
-  padding: 13px 30px;
-  cursor: pointer;
-  background-color: black;
-  color: white;
-  line-height: 1px;
-  margin-top: 20px;
-  border-radius: 3px;
-  border-style: none;
-}
-.googleLoginBtn {
-  height: 40px;
-  width: 100%;
-  font-size: 14px;
-  padding: 13px 30px;
-  cursor: pointer;
-  background-color: orangered;
-  color: white;
-  line-height: 1px;
-  margin-top: 10px;
-  margin-bottom: 12px;
-  border-radius: 3px;
-  border-style: none;
-}
-.SignUp > section > footer {
-  padding: 12px 16px;
-  text-align: center;
-}
-
-.loginLine {
-  color: #bcbcbc;
-  font-size: 15px;
-  margin-bottom: 35px;
-  a {
-    color: black;
-    text-decoration: underline;
-    cursor: pointer;
-  }
-}
-
-.noUser {
-  text-decoration: underline;
-}
-
-.SignUp.openModal {
-  display: flex;
-  align-items: center;
-  /* 팝업이 열릴때 스르륵 열리는 효과 */
-  animation: modal-bg-show 0.3s;
-}
-@keyframes modal-show {
-  from {
-    opacity: 0;
-    margin-top: -50px;
-  }
-  to {
-    opacity: 1;
-    margin-top: 0;
-  }
-}
-@keyframes modal-bg-show {
-  from {
-    opacity: 0;
-  }
-  to {
-    opacity: 1;
-  }
-}
